Another small cleanup for Ubuntug++ 4.3 warnings, a Makefile for unit tests, and MacOS fixes for OS tests adapt/os/ocpios/linux/src/OcpiOsTimer.cxx: fix MacOS timer precision to make unit test pass components/optical_flow_pyr_lk.rcc/optical_flow_pyr_lk.c: warnings core/container/ctests/bin/src/test_xml_worker_main.cxx: warnings core/container/hdl_container/impl/src/HdlContainer.cxx: warnings core/library/include/OcpiLibraryManager.h: warnings - new from g++ 4.3 core/local/util/fasttime/include/fasttime_private.h: warnings core/local/util/property/src/OcpiUtilDataTypes.cxx: warnings core/local/util/property/src/OcpiUtilProtocol.cxx: warnings core/sca/cf_util/vfs/src/OcpiCFUtilSCAFs.cxx: warnings core/sca/srpl/srpl/src/Cp289ExecutableDevice.cxx: warnings tests/facilities/test_os/tests/src/TestOcpiOsFileSystem.cxx: MacOS fix tools/cdk/examples/bestfit_test/Makefile: add "clean" target for when prerequisites NOT present tools/cdk/examples/canny/Makefile: ditto tools/cdk/examples/dds_ports/Makefile: ditto tools/cdk/examples/feature_detection/Makefile: ditto tools/cdk/examples/image_filtering/Makefile: ditto tools/cdk/examples/optical_flow/Makefile: ditto tools/cdk/examples/vsadd/Makefile: ditto tools/cdk/examples/dds_ports/specs/test.idl: reenable complex case tests/Makefile: for building and running tests

+++ components/optical_flow_pyr_lk.rcc/optical_flow_pyr_lk.c
@@ -130,8 +130,8 @@ void calcOpticalFlowPyrLK( OFState *myState,
   size_t derivJElemSize1 = sizeof(float);
   size_t derivIRows = H+2*winHeight;
   size_t derivJRows = H+2*winHeight;
-  size_t derivICols = W+2*winWidth;
-  size_t derivJCols = W+2*winWidth;
+  int derivICols = W+2*winWidth;
+  int derivJCols = W+2*winWidth;
 
 
   const char *derivIData = (const char *) myState->derivIBuf;
